move some folders around to follow other app structures
