+++ protocol-designer/src/file-data/__tests__/commandsSelectors.test.js
@@ -8,7 +8,7 @@ import {
   MULTI_PIPETTE,
   SOURCE_LABWARE,
   DEST_LABWARE,
-} from '../../step-generation/__tests__/fixtures'
+} from '../../step-generation/__fixtures__'
 
 import { getLabwareLiquidState, getRobotStateTimeline } from '../selectors'
 jest.mock('../../labware-defs/utils')
