+++ protocol-designer/src/step-generation/__fixtures__/commandFixtures.js
@@ -7,10 +7,7 @@ import type {
   TouchTipParams,
   Command,
 } from '@opentrons/shared-data/protocol/flowTypes/schemaV3'
-import type {
-  CommandsAndWarnings,
-  CommandCreatorErrorResponse,
-} from '../../types'
+import type { CommandsAndWarnings, CommandCreatorErrorResponse } from '../types'
 
 /** Used to wrap command creators in tests, effectively casting their results
  **  to normal response or error response
