+++ protocol-designer/src/step-generation/__tests__/dispenseUpdateLiquidState.test.js
@@ -7,7 +7,7 @@ import merge from 'lodash/merge'
 import omit from 'lodash/omit'
 import produce from 'immer'
 import { createEmptyLiquidState, createTipLiquidState } from '../utils'
-import { makeContext, DEFAULT_PIPETTE, SOURCE_LABWARE } from './fixtures'
+import { makeContext, DEFAULT_PIPETTE, SOURCE_LABWARE } from '../__fixtures__'
 
 import { dispenseUpdateLiquidState } from '../getNextRobotStateAndWarnings/dispenseUpdateLiquidState'
