+++ protocol-designer/src/step-generation/__tests__/distribute.test.js
@@ -17,7 +17,7 @@ import {
   pickUpTipHelper,
   dropTipHelper,
   ASPIRATE_OFFSET_FROM_BOTTOM_MM,
-} from './fixtures'
+} from '../__fixtures__'
 import { distribute } from '../commandCreators/compound/distribute'
 import type { DistributeArgs } from '../types'
