+++ protocol-designer/src/step-generation/__tests__/fixtureGeneration.test.js
@@ -1,6 +1,6 @@
 // @flow
 import { createEmptyLiquidState } from '../utils'
-import { makeContext, makeState } from './fixtures'
+import { makeContext, makeState } from '../__fixtures__'
 
 describe('snapshot tests', () => {
   test('makeContext', () => {
