+++ protocol-designer/src/step-generation/__tests__/forPickUpTip.test.js
@@ -1,12 +1,12 @@
 // @flow
 import merge from 'lodash/merge'
-import { makeImmutableStateUpdater } from './utils'
+import { makeImmutableStateUpdater } from '../__utils__'
 import {
   getInitialRobotStateStandard,
   makeContext,
   getTipColumn,
   DEFAULT_PIPETTE,
-} from './fixtures'
+} from '../__fixtures__'
 import { forPickUpTip as _forPickUpTip } from '../getNextRobotStateAndWarnings/forPickUpTip'
 
 import { dispenseUpdateLiquidState } from '../getNextRobotStateAndWarnings/dispenseUpdateLiquidState'
