+++ protocol-designer/src/step-generation/__tests__/transfer.test.js
@@ -15,7 +15,7 @@ import {
   dropTipHelper,
   ASPIRATE_OFFSET_FROM_BOTTOM_MM,
   DISPENSE_OFFSET_FROM_BOTTOM_MM,
-} from './fixtures'
+} from '../__fixtures__'
 import { transfer } from '../commandCreators/compound/transfer'
 
 const aspirateHelper = makeAspirateHelper()
