+++ protocol-designer/src/step-generation/__utils__/testMatchers.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-undef */
+
 // error of type exists somewhere in timeline errors
 export function expectTimelineError(errors, errorType) {
   expect(errors).toEqual(
